sharedlib/efficientdet/requirements.txt
#opencv-contrib-python==3.4.2.17
#opencv-python==3.4.2.17
numpy
Pillow>=6.2.2
tensorflow>=1.15
pytype
absl-py==0.9.0
git+https://github.com/cocodataset/cocoapi.git#subdirectory=PythonAPI
